requirements.txt
nltk==3.4.5
numpy==1.16.2
gensim==3.8.1
sklearn==0.0
tensorboard==1.12.2
tensorboardX==2.0
tensorflow-estimator==1.13.0
tensorflow-gpu==1.12.0
tensorflow-hub==0.4.0
cloudpickle==1.4.1
jsonpickle==1.4.1
xgboost==0.90

